Giving bootbox.prompt pre-selected value ability
I wanted to have the ability to allow for a pre-selected value when
using bootbox.prompt with a select box. This allows for that.

I also use SammyJS with a project and I made a small change that allows
BootBox to play nicely with it (and other similar JS routing libraries).